pyproject.toml
[project]
name = "function-semantic-kernel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-functions>=1.23.0",
    "python-dotenv>=1.1.1",
    "semantic-kernel>=1.35.3",
]
